+++ src/components/tab/TabGroup.story.vue
@@ -1,11 +1,173 @@
 <template>
     <Story title="Components/Tab/TabGroup">
-        Work in Progress
+        <Variant title="Default">
+            <TabGroup>
+                <TabPane label="Tab Item #1">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor 
+                    incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud 
+                    exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute 
+                    irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla 
+                    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui 
+                    officia deserunt mollit anim id est laborum.
+                </TabPane>
+
+                <TabPane label="Tab Item #2">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor 
+                    incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud 
+                    exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute 
+                    irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla 
+                    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui 
+                    officia deserunt mollit anim id est laborum.
+                </TabPane>
+                
+                <TabPane label="Tab Item #3">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor 
+                    incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud 
+                    exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute 
+                    irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla 
+                    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui 
+                    officia deserunt mollit anim id est laborum.
+                </TabPane>
+            </TabGroup>
+        </Variant>
+
+        <Variant title="Bordered">
+            <TabGroup bordered>
+                <TabPane label="Tab Item #1">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor 
+                    incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud 
+                    exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute 
+                    irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla 
+                    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui 
+                    officia deserunt mollit anim id est laborum.
+                </TabPane>
+
+                <TabPane label="Tab Item #2">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor 
+                    incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud 
+                    exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute 
+                    irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla 
+                    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui 
+                    officia deserunt mollit anim id est laborum.
+                </TabPane>
+                
+                <TabPane label="Tab Item #3">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor 
+                    incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud 
+                    exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute 
+                    irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla 
+                    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui 
+                    officia deserunt mollit anim id est laborum.
+                </TabPane>
+            </TabGroup>
+        </Variant>
+
+        <Variant title="Labels">
+            <div class="flex flex-col gap-8">
+                <TabGroup>
+                    <TabPane label="Inbox">
+                        <div class="p-8 text-center">
+                            Imagine your Inbox here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-inbox"><polyline points="22 12 16 12 14 15 10 15 8 12 2 12"/><path d="M5.45 5.11 2 12v6a2 2 0 0 0 2 2h16a2 2 0 0 0 2-2v-6l-3.45-6.89A2 2 0 0 0 16.76 4H7.24a2 2 0 0 0-1.79 1.11z"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                            <BadgeStd label="42" size="sm" />
+                        </template>
+                    </TabPane>
+
+                    <TabPane label="Drafts">
+                        <div class="p-8 text-center">
+                            Imagine some unsent draft messages here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-pencil-line"><path d="M12 20h9"/><path d="M16.5 3.5a2.12 2.12 0 0 1 3 3L7 19l-4 1 1-4Z"/><path d="m15 5 3 3"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                        </template>
+                    </TabPane>
+                    
+                    <TabPane label="Outbox">
+                        <div class="p-8 text-center">
+                            Imagine your Outbox here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-send"><path d="m22 2-7 20-4-9-9-4Z"/><path d="M22 2 11 13"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                        </template>
+                    </TabPane>
+                    
+                    <TabPane label="Trash">
+                        <div class="p-8 text-center">
+                            Imagine some trashed messages here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-trash-2"><path d="M3 6h18"/><path d="M19 6v14c0 1-1 2-2 2H7c-1 0-2-1-2-2V6"/><path d="M8 6V4c0-1 1-2 2-2h4c1 0 2 1 2 2v2"/><line x1="10" x2="10" y1="11" y2="17"/><line x1="14" x2="14" y1="11" y2="17"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                            <BadgeStd label="8" size="sm" color="danger" />
+                        </template>
+                    </TabPane>
+                </TabGroup>
+                
+                <TabGroup bordered>
+                    <TabPane label="Inbox">
+                        <div class="p-8 text-center">
+                            Imagine your Inbox here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-inbox"><polyline points="22 12 16 12 14 15 10 15 8 12 2 12"/><path d="M5.45 5.11 2 12v6a2 2 0 0 0 2 2h16a2 2 0 0 0 2-2v-6l-3.45-6.89A2 2 0 0 0 16.76 4H7.24a2 2 0 0 0-1.79 1.11z"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                            <BadgeStd label="42" size="sm" />
+                        </template>
+                    </TabPane>
+
+                    <TabPane label="Drafts">
+                        <div class="p-8 text-center">
+                            Imagine some unsent draft messages here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-pencil-line"><path d="M12 20h9"/><path d="M16.5 3.5a2.12 2.12 0 0 1 3 3L7 19l-4 1 1-4Z"/><path d="m15 5 3 3"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                        </template>
+                    </TabPane>
+                    
+                    <TabPane label="Outbox">
+                        <div class="p-8 text-center">
+                            Imagine your Outbox here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-send"><path d="m22 2-7 20-4-9-9-4Z"/><path d="M22 2 11 13"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                        </template>
+                    </TabPane>
+                    
+                    <TabPane label="Trash">
+                        <div class="p-8 text-center">
+                            Imagine some trashed messages here
+                        </div>
+
+                        <template #header="{ label }">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-trash-2"><path d="M3 6h18"/><path d="M19 6v14c0 1-1 2-2 2H7c-1 0-2-1-2-2V6"/><path d="M8 6V4c0-1 1-2 2-2h4c1 0 2 1 2 2v2"/><line x1="10" x2="10" y1="11" y2="17"/><line x1="14" x2="14" y1="11" y2="17"/></svg>
+                            <span class="tab-label">{{ label }}</span>
+                            <BadgeStd label="8" size="sm" color="danger" />
+                        </template>
+                    </TabPane>
+                </TabGroup>
+            </div>
+        </Variant>
     </Story>
 </template>
 
 <script lang="ts" setup>
-
+import BadgeStd from '../badge/BadgeStd.vue';
+import TabGroup from './TabGroup.vue';
+import TabPane from './TabPane.vue';
 </script>
 
 <docs lang="md">